Feat(tests): end-to-end test of connector happy path (deposit and withdraw) (#15)

+++ aurora-locker/src/Codec.sol
@@ -82,7 +82,7 @@ library Codec {
 
     /// Decode promise result using borsh.
     function decodePromiseResult(Borsh.Data memory data) public pure returns (PromiseResult memory result) {
-        result.status = PromiseResultStatus(uint8(data.decodeU8()));
+        result.status = PromiseResultStatus(data.decodeU8());
         if (result.status == PromiseResultStatus.Successful) {
             result.output = data.decodeBytes();
         }
